Change module names
The split between RFC1524 and Mailcap didn't make much sense to begin
with. Furthermore RFC's change, are superseded by newer RFCs and naming
them after it is flawed.

Fixes https://github.com/purebred-mua/purebred-mailcap/issues/6
