+++ purebred-mailcap.cabal
@@ -25,9 +25,8 @@ extra-source-files:  CHANGELOG.md
 
 library
     exposed-modules:  Data.Mailcap
-                    , Data.RFC1524
-                    , Data.RFC1524.Internal
-                    , Data.RFC1524.ViewCommand
+                    , Data.Mailcap.Internal
+                    , Data.Mailcap.ViewCommand
 
     build-depends:    base >= 4.11 && < 5
                     , attoparsec
